backend/requirements.txt
# Compatible with Python 3.13
Django==5.1.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.3
django-fsm==2.8.1
channels==4.1.0
channels-redis==4.2.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.11
python-decouple==3.8
Pillow==11.1.0
drf-spectacular==0.27.2
boto3==1.34.0
razorpay==1.4.1
gunicorn==23.0.0
daphne==4.1.2
sentry-sdk==2.20.0
python-magic==0.4.27
